chore: move traversal utils to service
